requirements.txt
backcall==0.1.0
backports.csv==1.0.7
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.2
django-admin-interface==0.10.5
django-colorfield==0.1.15
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-import-export==1.2.0
dnspython==1.16.0
et-xmlfile==1.0.1
freeze==1.0.10
future==0.17.1
gunicorn==19.9.0
ipython==7.5.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.0
mysql-connector==2.2.9
mysqlclient==1.4.2.post1
numpy==1.16.4
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.24.2
parso==0.5.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
Pygments==2.4.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
pywatchman==1.4.1
PyYAML==5.1.1
reportlab==3.5.23
scipy==1.3.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==4.1.3
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
